Eliminate the posts_admin view because it's shit
posts_view is up next. We will also be getting rid of Database.php, as
the idea of a centralised class for every database action was silly (but
at least it wasn't mine).

From now on, SQL queries will be treated as "first-class citizens" and
gradually moved to services/views/whatever as appropriate. This makes
sense because the same SQL queries are rarely used in more than one
place.